pass select element down into component; move show back up to tableview
