+++ src/DashboardUI/src/components/home-page/TableView.tsx
@@ -38,7 +38,8 @@ function TableView() {
             </Nav.Item>
           </Nav>
           <Tab.Content>
-            <ChartsController activeTab={activeTab} show={show} />
+            {show && <ChartsController activeTab={activeTab} />}
+
             <Tab.Pane eventKey="dataTable">
               {activeTab === 'dataTable' && show && <AnalyticsDataTable></AnalyticsDataTable>}
             </Tab.Pane>
